requirements.txt
# Core API
fastapi==0.115.0
uvicorn==0.30.0
pydantic==2.8.2

# Database
sqlalchemy==2.0.30
psycopg2-binary==2.9.9

# Env
python-dotenv==1.0.1

# AI / LLM
groq==0.9.0
langgraph==0.2.3
langchain-core==0.2.38

# Utilities
python-multipart==0.0.9

httpx==0.27.0